+++ src/main/java/org/dialog/hack/repository/EmployeeSupervisorRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by sasini on 5/24/17.
  */
-public interface EmployeeSupervisorRepository extends CrudRepository<EmployeeSupervisor, Integer> {
+public interface EmployeeSupervisorRepository extends CrudRepository<EmployeeSupervisor, Long> {
 
 }
